test(EuiDataGridCell): move EuiDataGrid cell rendering jest tests to a separate file
