+++ packages/eui/src/components/datagrid/data_grid.test.tsx
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { mount, ReactWrapper } from 'enzyme';
 import { act, fireEvent } from '@testing-library/react';
 import { findTestSubject, requiredProps } from '../../test';
@@ -467,82 +467,6 @@ describe('EuiDataGrid', () => {
       renderCellValue={() => null}
     />
   );
-
-  describe('cell rendering', () => {
-    it('supports hooks', () => {
-      const RenderCellValueWithHooks: RenderCellValue = ({
-        rowIndex,
-        columnId,
-      }) => {
-        const [value] = useState(`Hello, Row ${rowIndex}-${columnId}!`);
-        return <span>{value}</span>;
-      };
-      const component = mount(
-        <EuiDataGrid
-          aria-label="test"
-          columns={[{ id: 'Column 1' }, { id: 'Column 2' }]}
-          columnVisibility={{
-            visibleColumns: ['Column 1', 'Column 2'],
-            setVisibleColumns: () => {},
-          }}
-          rowCount={2}
-          renderCellValue={RenderCellValueWithHooks}
-        />
-      );
-      expect(extractGridData(component)).toMatchInlineSnapshot(`
-        [
-          [
-            "Column 1",
-            "Column 2",
-          ],
-          [
-            "Hello, Row 0-Column 1!",
-            "Hello, Row 0-Column 2!",
-          ],
-          [
-            "Hello, Row 1-Column 1!",
-            "Hello, Row 1-Column 2!",
-          ],
-        ]
-      `);
-    });
-
-    it('passes `cellContext` as props to the renderCellValue component', () => {
-      const dataGridProps = {
-        'aria-label': 'test',
-        columns: [{ id: 'Column' }],
-        columnVisibility: {
-          visibleColumns: ['Column'],
-          setVisibleColumns: () => {},
-        },
-        rowCount: 1,
-      };
-
-      const RenderCellValueWithContext: RenderCellValue = ({ someContext }) => (
-        <div data-test-subj="renderedCell">
-          {someContext ? 'hello' : 'world'}
-        </div>
-      );
-
-      const { getByTestSubject, rerender } = render(
-        <EuiDataGrid
-          {...dataGridProps}
-          renderCellValue={RenderCellValueWithContext}
-          cellContext={{ someContext: true }}
-        />
-      );
-      expect(getByTestSubject('renderedCell')).toHaveTextContent('hello');
-
-      rerender(
-        <EuiDataGrid
-          {...dataGridProps}
-          renderCellValue={RenderCellValueWithContext}
-          cellContext={{ someContext: false }}
-        />
-      );
-      expect(getByTestSubject('renderedCell')).toHaveTextContent('world');
-    });
-  });
 
   describe('pagination', () => {
     it('renders', () => {
